+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- updated variable definition *variables.tf*
+
 ### Fixed
 
 - fix ADB *db_name* be be unique